+++ server/src/main/java/org/opensearch/identity/package-info.java
@@ -1,0 +1,17 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ *
+ * Modifications Copyright OpenSearch Contributors. See
+ * GitHub history for details.
+ */
+
+/**
+ * Holds classes related to the identity owners of request to/from extensions
+ *
+ * @opensearch.internal
+ */
+package org.opensearch.identity;